Fix cases where gethostname() returns a name that doesn't work for client
We currently rely on gethostname() to generate the links from the web
control homepage to the camera streams - unfortunately if DNS isn't
set up well the hostname won't be one the web browser can resolve
(often the case on raspberry pi's, where the hostname is set to
"raspberrypi" by default and users access them via IP address).

To cater for these situations we look at the "Host:" header in the
HTTP headers that came from the browser (which indicate which host the
browser has been asked to connect to) and use that to generate the
hostname to use in the links.

If there's no Host: header (unlikely these days) we default back to
the old behaviour.