Fix for crashes in rtsp netcam resize code
The current code resulted in the motion_loop thread accessing (and
writing to) variables that should only be accessed in the
netcam_handler_loop thread, meaning we could see crashes.

We also introduce the necessary mutex around the remaining variables
in netcam_read_rtsp_image so:

a) we can be sure the netcam won't mess with the image buffer whilst
we're reading it
b) that any new image in 'latest' has actually been passed across the
threads and isn't lying in a register/CPU cache somewhere.

As per https://github.com/Motion-Project/motion/issues/224 the probem
code has been moved onto the netcam thread where it should be safe.

Nominally this may be less efficient (as we will resize every frame,
but motion_loop may process only some of those frames); in the long term
it may well be better to have the netcam_handler_loop send a full "image"
structure to the motion_loop thread (ie. including size and format
information), which the motion_loop thread would then be able to safely
resize.

Fixes #207
Fixes #224