pts change for libav
The presentation time stamp(pts) indicates the time which the movie
player should display a particular frame.  In the previous changes
to the pts, we set the time basis to be in microseconds by specifying
the denomintor to be equal to 1000.  This allows for additional precision
when we assign the pts when writing the frame.  For the libav variant
the time basis assignment is ignored and instead the framerate is
used.  The use of the framerate causes problems when it is small.
At 2fps, time intervals of 1 to 499 microseconds round to a pts
of zero, 500 to 1499 microseconds round to pts 1, etc.  This
commit adds a variable to keep track of the last pts used in writing
the frame.  If the newly calculated pts is not greater than the previous
pts, we just increment the pts by 1 which would be the full 1/framerate
increment.  This could result visually twitchy videos at low framerates.
It could also have a compounding affect of eliminating the handling of
variable framerates and just fallback to a constant framerate video.