Add mask_privacy option
This commit adds the functionality to completely mask out
user specified areas of the images via a pgm file.   The
masking is applied only to the y component of the image
From reviewing the resulting image, it does not appear that
it is necessary to apply the mask to the cr or cb components.
It may be desired to apply the mask to these components in the
future to have a true digital mask rather than just a visual
mask.  As indicated in the code comments, the resulting blocked
section appear as white in the images since this facilitates
quicker processing.