Eliminate pauses in movies during events
The PTS is based upon the actual time associated with the image
when we have the motion->no motion->motion situation, the second
motion is setting the time from that image and calculating the
duration from the initial start of the movie.  During this entire
sequence, there is a ffmpeg_output defined.  This commit resets the
starting time after there is a period of no motion.