v4l update
This update removes the components for v4l and implements only
the v4l2 and BSD option.  This commit also consolidates some
of the pre-processor directives and only uses the WITHOUT_V4L2
instead of the multiple.  The checks for the HAVE_LINUX_VIDEODEV2_H
were removed from the code because the WITHOUT_V4L2 is actually
set based upon the existance of the HAVE_LINUX_VIDEODEV2_H.
Finally, this renames some files to be more consistent.