Privacy Mask
This commit fully implements the privacy mask on all color components of the image.

Since activity could still be detected with just the cr/cb this change addresses
putting the mask on these color components in addition to just the y component.

It also changes the mask to black since we have to manually assign the cr/cb
to x80 to have it reflect a acceptable color for the masked area.