Improvements to the `camera_dir` config option
- The `camera_dir` option was always being set to a default path
which was missing a 'motion/' compared to the installed config
file default. This caused it to always check that folder and
upon a Web config-write, to always write out as enabled.
We now set it NULL like most other options and read the config
value into it, if enabled.
- Upon a Web config-write if the option is disabled, we now write
out the installed default path with it (remaining disabled).
- Improved the guide and config help to be more clear of how the
option works.