Thread Safety Change for MJPG V4L2
This commit revises the jpeg utility for thread safety by moving the
declares for the buf0-2 and chr1-2 variables into the associated
functions rather than being at the module level hidden in the middle
of the code.

For the decode_jpeg_gray_raw this routine appears to have been copied
from the decode_jpeg_raw but when it was initially created.  When it was
copied, it did not however remove all the code that was related to the
Cr and Cb segments which do not exist in grey images.  Since the compiler
now reports these as unused, this commit removes them to clean up the code.