Rewrite jpegutils.c
The previous version of jpegutils.c was performing the downsampling of images manually
This commit rewrites the decompression to take advantage of the built in downsampling
and image color space conversion that is part of the JPEG libaries.

This function is currently only called from the vid_mjpegtoyuv420p routine.