use the same timestamp for FILECREATE and FILECLOSE events
Motion uses a timestamp for the "Conversion Specifiers" which can be
used when creating the filenames for output filenames.
Propagate this timestamp to the FILECREATE and FILECLOSE events, so the
timestamp available for the "Conversion Specifiers" for SQL queries
matches the timestamp used in filenames.

To have currenttime_tv available, the FILE_CREATE event call is moved from picture.c:put_picture() to its calling functions in event.c.