Fix webu_process_action for a default single camera config
When using the single 'motion.conf' file for a single camera setup,
the webu actions are skipping over the only thread due to assuming
that thread '0' is just motion and all cameras are subsequent.
This leaves the webu actions useless to the default single camera,
single config setup.

Fix this by checking that more than one camera thread exists to use
the thread loop, using the given thread number ('0') otherwise.