systemd unit file enhancements and cleanup (#907)
* Remove unused PIDFile setting

Since the default config doesn't create a PID file, and they're not
needed for systemd unit files which aren't forking (this unit file
runs motion as a 'simple' service), this setting can be removed in the
default config.

* Add comment about the systemd unit file, including license info.

* Expand description, add man page reference.

* systemd: read /etc/default/motion for compatibility with init script.

Allow the user to specify environment variables to be inherited by
motion when started via native systemd service.  Noop if file does not
exist.  Useful to pass env files to scripts etc. which are called by
motion (e.g. with on_event_start etc.).

* Comment discarding of stderr.

* Add reload functionality to systemd unit file.

* Add restart behaviour to systemd unit file.

* Set motion to run as user "motion" for consistency with init scripts

* Honor start_motion_daemon setting in /etc/default/motion

For existing installations which have disabled auto-start of the daemon
by setting start_motion_daemon != yes in /etc/default/motion, check this
variable as a prerequisite for systemd to start motion.

* systemd: Document some security related options, don't enable by default.